Reuse code for handling sink response
