+++ data-plane/dispatcher/src/main/java/dev/knative/eventing/kafka/broker/dispatcher/impl/ResponseToKafkaTopicHandler.java
@@ -17,29 +17,20 @@ package dev.knative.eventing.kafka.broker.dispatcher.impl;
 
 import dev.knative.eventing.kafka.broker.core.AsyncCloseable;
 import dev.knative.eventing.kafka.broker.core.metrics.Metrics;
-import dev.knative.eventing.kafka.broker.core.tracing.TracingSpan;
 import dev.knative.eventing.kafka.broker.dispatcher.ResponseHandler;
 import io.cloudevents.CloudEvent;
-import io.cloudevents.http.vertx.VertxMessageFactory;
 import io.vertx.core.CompositeFuture;
 import io.vertx.core.Future;
-import io.vertx.core.buffer.Buffer;
-import io.vertx.ext.web.client.HttpResponse;
 import io.vertx.kafka.client.producer.KafkaProducer;
 import io.vertx.kafka.client.producer.KafkaProducerRecord;
-import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.Objects;
 
-import static dev.knative.eventing.kafka.broker.core.utils.Logging.keyValue;
-
 /**
  * This class implements a {@link ResponseHandler} that will convert the sink response into a {@link CloudEvent} and push it to a Kafka topic.
  */
-public final class ResponseToKafkaTopicHandler implements ResponseHandler {
-
-  private static final Logger logger = LoggerFactory.getLogger(ResponseToKafkaTopicHandler.class);
+public final class ResponseToKafkaTopicHandler extends BaseResponseHandler implements ResponseHandler {
 
   private final String topic;
   private final KafkaProducer<String, CloudEvent> producer;
@@ -52,6 +43,7 @@ public final class ResponseToKafkaTopicHandler implements ResponseHandler {
    * @param topic    topic to produce records.
    */
   public ResponseToKafkaTopicHandler(final KafkaProducer<String, CloudEvent> producer, final String topic) {
+    super(LoggerFactory.getLogger(ResponseToKafkaTopicHandler.class));
 
     Objects.requireNonNull(topic, "provide topic");
     Objects.requireNonNull(producer, "provide producer");
@@ -61,53 +53,11 @@ public final class ResponseToKafkaTopicHandler implements ResponseHandler {
     this.producerMeterBinder = Metrics.register(this.producer.unwrap());
   }
 
-  /**
-   * Handle the given response.
-   *
-   * @param response response to handle
-   * @return a succeeded or failed future.
-   */
   @Override
-  public Future<Void> handle(final HttpResponse<Buffer> response) {
-    CloudEvent event;
-
-    try {
-      event = VertxMessageFactory.createReader(response).toEvent();
-    } catch (final Exception ex) {
-      if (maybeIsNotEvent(response)) {
-        if (logger.isDebugEnabled()) {
-          logger.debug(
-            "Response is not recognized as event, discarding it {} {} {}",
-            keyValue("response", response),
-            keyValue("response.body", response == null || response.body() == null ? "null" : response.body()),
-            keyValue("response.body.len",
-              response == null || response.body() == null ? "null" : response.body().length())
-          );
-        }
-        return Future.succeededFuture();
-      }
-
-      // When the sink returns a malformed event we return a failed future to avoid committing the message to Kafka.
-      return Future.failedFuture(
-        new IllegalStateException("Unable to decode response: unknown encoding and non empty response", ex)
-      );
-    }
-
-    if (event == null) {
-      return Future.failedFuture(new IllegalArgumentException("event cannot be null"));
-    }
-
-    TracingSpan.decorateCurrentWithEvent(event);
-
+  protected Future<Void> doHandleEvent(CloudEvent event) {
     return producer
       .send(KafkaProducerRecord.create(topic, event))
       .mapEmpty();
-  }
-
-  private static boolean maybeIsNotEvent(final HttpResponse<Buffer> response) {
-    // This checks whether there is something in the body or not, though binary events can contain only headers and they
-    // are valid Cloud Events.
-    return response == null || response.body() == null || response.body().length() <= 0;
   }
 
   @Override
